Added DuplicateCustomerException
DuplicateCustomerException is to be thrown when a customer with the exact same first, last, email, and phone number as an existing entry is tried to be created.

+++ MainService/src/main/java/com/eklof/CustomersAPI/services/CustomerService.java
@@ -1,10 +1,13 @@
 package com.eklof.CustomersAPI.services;
 
+import com.eklof.CustomersAPI.exceptions.DuplicateCustomerException;
 import com.eklof.CustomersAPI.exceptions.NoSuchCustomerException;
 import com.eklof.CustomersAPI.models.Customer;
 import com.eklof.CustomersAPI.repositories.CustomerRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.function.Consumer;
 
 @Service
 public class CustomerService implements ICustomerService {
@@ -33,12 +36,30 @@ public class CustomerService implements ICustomerService {
         customer.setLastName(last);
         customer.setEmail(email);
         customer.setPhoneNumber(phoneNumber);
+        Iterable<Customer> customers = customerRepository.findAll();
+        customers.forEach(new Consumer<Customer>() {
+
+            @Override
+            public void accept(Customer cust) {
+                Boolean duplicate = false;
+                if (cust.getFirstName().equals(first) && cust.getLastName().equals(last) && cust.getEmail().equals(email) && cust.getPhoneNumber().equals(phoneNumber)) {
+                    duplicate = true;
+                }
+                if (duplicate) {
+                    throw new DuplicateCustomerException();
+                }
+            }
+
+        });
         customerRepository.save(customer);
         return customer;
     }
 
     @Override
     public void deleteCustomer(Integer id) {
+        if (!customerRepository.existsById(id)) {
+            throw new NoSuchCustomerException(id);
+        }
         customerRepository.deleteById(id);
     }
 }
